Modify the routes structure
The structure of application routing was flawed. This is a try to fix
that. The main goal was to keep the app.js minimal, thus it only
includes the routes from the subdirectory. The index.coffee file is kind
of a "proxy" that is included from the app.js and then includes the
actual routes from other files.

Each file in the routes directory should contain the routes related to
same object. For example one file for the main view, one file to
authentication etc.